+++ CHANGELOG.md
@@ -16,6 +16,7 @@ all of them.
  - Add extension `VK_KHR_external_memory` (revision 1)
  - Add extension `VK_KHR_external_memory_win32` (revision 1)
  - Add extension `VK_KHR_external_memory_fd` (revision 1)
+ - Add extension `VK_KHR_win32_keyed_mutex` (revision 1)
 
 ### Breaking changes